==> 2.Population&Vaccination.R <==

## Create Pemba maps
source("CreateMaps.R")

## Create Pemba grids
source("CreateGrids.R")

## Prepare vaccination data
source("matchVaxData.R")

## Prepare transect data
source("matchTransectData.R")

## Prepare government dog census data
source("matchGovCensusData.R")

## Prepare 2012 NBS human census data
source("matchNBS2012Village.R")

## Estimate human and dog populations
source("CreatePop.R")

## Estimate vaccination coverage
source("CreateVacc.R")
